Replacevars get updated on search appearance pages when separator gets updated
